secret-backend: Allow override the backend also under Flatpak/Snap
This can help applications, which rely on certain behavior, like when
multiple processes are accessing the secrets and expect that secrets
saved by one process can be read by another process.

It also adds consistency, to be able to change the backend with
the host system libsecret and sandboxed libsecret.

Related to https://gitlab.gnome.org/GNOME/libsecret/-/issues/62